Backend/requirements.txt
faker
flask
joblib
matplotlib
numpy
paho-mqtt
pandas
python-dotenv
pyyaml
scikit-learn
